requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
assemblyai==0.17.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.3.23
clarifai==9.8.0
clarifai-grpc==9.8.0
click==8.1.7
clickhouse-connect==0.6.11
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.5.14
duckdb==0.8.1
exceptiongroup==1.1.3
fastapi==0.99.1
filelock==3.12.3
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.58.0
h11==0.14.0
hnswlib==0.7.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
langchain==0.0.190
langsmith==0.0.35
llama-index==0.6.11
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
onnxruntime==1.15.1
openai==0.28.0
openapi-schema-pydantic==1.2.4
overrides==7.4.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
posthog==3.0.2
protobuf==4.24.3
pulsar-client==3.3.0
pydantic==1.10.12
pydantic_core==2.6.3
Pygments==2.16.1
PyPika==0.48.9
pyreadline3==3.4.1
python-certifi-win32==1.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-rapidjson==1.10
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.29.0
rich==13.4.2
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools-scm==7.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.1
tritonclient==2.34.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
youtube-dl==2021.12.17
zipp==3.16.2
zstandard==0.21.0
